+++ packages/assets/__tests__/url.test.js
@@ -3,11 +3,7 @@ import { createConfig, match } from '@webpack-blocks/core'
 import { url } from '../lib/index'
 
 test('url() works with match', t => {
-  const config = createConfig({}, [
-    match('*.jpg', [
-      url()
-    ])
-  ])
+  const config = createConfig({}, [match('*.jpg', [url()])])
 
   t.deepEqual(config.module.rules, [
     {
@@ -40,45 +36,6 @@ test('url() works with options and match()', t => {
           loader: 'url-loader',
           options: {
             mimetype: 'image/gif'
-          }
-        }
-      ]
-    }
-  ])
-})
-
-test('deprecated url(<fileType>) still works', t => {
-  const config = createConfig({}, [
-    url('image')
-  ])
-
-  t.deepEqual(config.module.rules, [
-    {
-      test: /\.(gif|ico|jpg|jpeg|png|svg|webp)$/,
-      use: [
-        {
-          loader: 'url-loader',
-          options: {}
-        }
-      ]
-    }
-  ])
-})
-
-test('deprecated url(<fileType>, { exclude }) still works', t => {
-  const config = createConfig({}, [
-    url('image', { exclude: /node_modules/ })
-  ])
-
-  t.deepEqual(config.module.rules, [
-    {
-      test: /\.(gif|ico|jpg|jpeg|png|svg|webp)$/,
-      exclude: /node_modules/,
-      use: [
-        {
-          loader: 'url-loader',
-          options: {
-            exclude: /node_modules/
           }
         }
       ]