+++ packages/sample-app/webpack.config.babel.js
@@ -3,11 +3,11 @@ const {
 
   // Feature blocks
   addPlugins,
-  defineConstants,
   entryPoint,
   env,
   group,
   performance,
+  setMode,
   setOutput,
   sourceMaps,
 
@@ -16,55 +16,42 @@ const {
   css,
   devServer,
   extractText,
-  typescript,
-  uglify
+  typescript
 } = require('webpack-blocks')
 
 const HtmlWebpackPlugin = require('html-webpack-plugin')
-const webpack = require('webpack')
+const path = require('path')
 
-const developmentConfig = () => group([
-  sourceMaps(),
-  devServer({
-    proxy: {
-      '/api/*': { target: 'http://localhost:4000' }
-    }
-  }),
-  performance({
-    // Increase performance budget thresholds for development mode
-    maxAssetSize: 1500000,
-    maxEntrypointSize: 1500000
-  })
-])
-
-const productionConfig = () => group([
-  extractText(),
-  uglify(),
-  addPlugins([
-    new webpack.LoaderOptionsPlugin({
-      minimize: true,
-      debug: false
-    })
+const developmentConfig = () =>
+  group([
+    sourceMaps(),
+    devServer({
+      proxy: {
+        '/api/*': { target: 'http://localhost:4000' }
+      }
+    }),
+    performance({
+      // Increase performance budget thresholds for development mode
+      maxAssetSize: 1500000,
+      maxEntrypointSize: 1500000
+    }),
+    css.modules()
   ])
-])
+
+const productionConfig = () =>
+  group([css.modules({ minimize: true }), extractText('css/[name].[contenthash:hex:8].css')])
 
 module.exports = createConfig([
+  setMode(process.env.NODE_ENV || 'development'),
   babel(),
-  typescript(),
-  css.modules(),
+  typescript({ configFileName: path.resolve(__dirname, './tsconfig.json') }),
   addPlugins([
     new HtmlWebpackPlugin({
       inject: true,
       template: './index.html'
     })
   ]),
-  defineConstants({
-    'process.env.NODE_ENV': process.env.NODE_ENV || 'development'
-  }),
-  env('development', [
-    entryPoint('./src/index.dev.js'),
-    developmentConfig()
-  ]),
+  env('development', [entryPoint('./src/index.dev.js'), developmentConfig()]),
   env('production', [
     entryPoint('./src/index.js'),
     setOutput('./build/bundle.js'),