+++ packages/webpack/__e2e-fixtures__/minimal/webpack.config.js
@@ -1,14 +1,11 @@
-const { createConfig, customConfig, entryPoint, setOutput } = require('../../index')
+const { createConfig, customConfig, entryPoint, setMode, setOutput } = require('../../index')
 
 const path = require('path')
 
 module.exports = createConfig([
-  entryPoint(
-    path.join(__dirname, 'app.js')
-  ),
-  setOutput(
-    path.join(__dirname, 'build/bundle.js')
-  ),
+  setMode('development'),
+  entryPoint(path.join(__dirname, 'app.js')),
+  setOutput(path.join(__dirname, 'build/bundle.js')),
   customConfig({
     output: {
       // Created bundle will be a module instead of a stand-alone self-running bundle