+++ packages/webpack/__e2e-fixtures__/postcss-sass-sourcemaps/webpack.config.js
@@ -1,4 +1,4 @@
-const { createConfig, entryPoint, setOutput } = require('../../index')
+const { createConfig, entryPoint, setMode, setOutput } = require('../../index')
 
 const path = require('path')
 const postcss = require('@webpack-blocks/postcss')
@@ -6,16 +6,11 @@ const precss = require('precss')
 const sass = require('@webpack-blocks/sass')
 
 module.exports = createConfig([
-  entryPoint(
-    path.join(__dirname, 'styles.sass')
-  ),
-  setOutput(
-    path.join(__dirname, 'build/bundle.js')
-  ),
+  setMode('development'),
+  entryPoint(path.join(__dirname, 'styles.sass')),
+  setOutput(path.join(__dirname, 'build/bundle.js')),
   postcss({
-    plugins: [
-      precss
-    ]
+    plugins: [precss]
   }),
   sass({ indentSyntax: true, sourceMap: true })
 ])