+++ packages/webpack/__e2e-fixtures__/sass-extract-text/webpack.config.js
@@ -1,18 +1,12 @@
-const { createConfig, entryPoint, match, setOutput } = require('../../index')
+const { createConfig, entryPoint, match, setMode, setOutput } = require('../../index')
 
 const path = require('path')
 const extractText = require('@webpack-blocks/extract-text')
 const sass = require('@webpack-blocks/sass')
 
 module.exports = createConfig([
-  entryPoint(
-    path.join(__dirname, 'app.js')
-  ),
-  setOutput(
-    path.join(__dirname, 'build/bundle.js')
-  ),
-  match('*.sass', [
-    sass({ indentSyntax: true }),
-    extractText('styles.css')
-  ])
+  setMode('development'),
+  entryPoint(path.join(__dirname, 'app.js')),
+  setOutput(path.join(__dirname, 'build/bundle.js')),
+  match('*.sass', [sass({ indentSyntax: true }), extractText('styles.css')])
 ])