+++ packages/assets/CHANGELOG.md
@@ -1,10 +1,14 @@
 # @webpack-blocks/assets - Changelog
 
+## Next Release
+
+- Remove deprecated `fileType` API ([#260](https://github.com/andywer/webpack-blocks/issues/260))
+- Support for webpack 4 ([#261](https://github.com/andywer/webpack-blocks/pull/261))
+
 ## 1.0.0-rc
 
-- Added a `styleLoader` option to `css()` and `css.modules()` blocks.
-With it it's now possible to pass options to the `style-loader` or
-remove it from loaders completely.
+- Added a `styleLoader` option to `css()` and `css.modules()` blocks. With it it's now possible to
+  pass options to the `style-loader` or remove it from loaders completely.
 
 ## 1.0.0-beta.2